Pass the syscall number to capsicum permission-denied signals
Summary: The syscall number is stored in the same register as the
syscall return on amd64 (and possibly other architectures) and so it is
impossible to recover in the signal handler after the call has returned.
This small tweak delivers it in the `si_value` field of the signal,
which is sufficient to catch capability violations and emulate them with
a call to a more-privileged process in the signal handler.

Reviewers: #capsicum, #manpages, manu, kib

Reviewed By: #manpages, kib

Subscribers: jhb, bcr, andrew, jhibbits, imp, lwhsu, emaste, jonathan, kib

Differential Revision: https://reviews.freebsd.org/D29185