Update snmalloc import to FreeBSD
Snmalloc is a modern high-performance memory allocator that is developed
by Microsoft Research.  It is also being used to as the testbed for a
lot of security extensions, including CHERI temporal safety.

This uses snmalloc's memcpy unless building with
`WITHOUT_SNMALLOC_MEMCPY` set.

The build has not been tested on all architectures and some may be
broken.

In our testing (limited number of microarchitectures), snmalloc's C++
memcpy is faster than the assembly versions shipped by FreeBSD for
PowerPC and amd64.  It can be tuned for different platforms and
microarchitectures with some fairly usable hooks and we'd be very happy
to take more variants upstream.

The memcpy can perform bounds checks on any heap source or destination.
This is controlled via the LIBC_BOUNDS_CHECKS environment variable,
which has three supported values:

0: No checks, I don't care about security.
1: Checks on destination, I care about integrity.
2: Checks on source and destination, I care about integrity and
   confidentiality.

Each layer of checking adds a small amount of overhead, though in our
testing the version with checks on the destination performed about as
well as the assembly version currently in libc.

An ifunc resolver picks the correct implementation the first time that
memcpy is called.  rtld needs to work without ifuncs, so now has its own
simple memcpy.  This is probably not ideal and we could also use the
snmalloc memcpy there without the ifunc or bounds checks.

The checks can easily be added to other libc functions, for example
sprintf and so on.

Without MALLOC_PRODUCTION, all of the other security checks are also
compiled in and the bounds checks will provide an error message